Added router and Dev UI
Dev UI is for navigating to different pages easily, also shows a mobile view port on the side.

This WILL mess with CSS values such as 'vh' and 'vw'.

If you are working with those values turn DevMode off in the app file. And you should be able to develop like before